materials/requirements.txt
pandas
azure-storage-file-datalake
azure-identity
python-dotenv